ml_project/requirements.txt
click==7.1.2
flake8==3.7.9
hydra-core==1.0.6
numpy==1.20.2
omegaconf==2.0.6
pandas==1.2.3
pylint==2.8.2
pytest==6.2.4
pytest-cov==2.11.1
PyYAML==5.4.1
scikit-learn==0.24.1
